update the routes wording with create and add session check for modify button in movie page
