+++ app/movies/create/layout.tsx
@@ -6,7 +6,7 @@ export const metadata: Metadata = {
   description:
     "Ajoutez et référencez vos films facilement en utilisant Movie Diary",
   alternates: {
-    canonical: `/movies/contribute`,
+    canonical: `/movies/create`,
   },
 };
