+++ app/profile/page.tsx
@@ -14,7 +14,7 @@ export default async function ProfilePage() {
         <p className="text-xl text-rose-500 font-semibold mb-4">Contribuer</p>
 
         <Link
-          href="/movies/contribute"
+          href="/movies/create"
           className="bg-gradient-to-r from-rose-500 to-red-500 text-white px-4 py-2 rounded-md hover:from-rose-600  hover:to-red-600"
         >
           Ajouter un film
